api/requirements.txt
fastapi
uvicorn
pandas
alpaca-py
python-dotenv
